7_stack-analyzer/StackSnap/pyproject.toml
[project]
name = "stacksnap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "manas-099", email = "manaspatra481@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",

    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
stacksnap = "stacksnap:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
